refactor: modularize code
