+++ lua/yasl/highlights.lua
@@ -1,6 +1,6 @@
-local M = {}
-
 local fallback_color = vim.api.nvim_get_hl(0, { name = "StatusLine" })
+
+local M = {}
 
 function M.set_hl(name, highlight)
 	vim.api.nvim_set_hl(0, name, vim.F.if_nil(highlight, fallback_color))
